GadgetField: use an external field validator.
GadgetField doesn't know how to validate/store REQUEST form data.
Use an external field to handle this task.
The external field must be located in the same form, and can be put in the hidden group if not rendered.

Add a test with a stringfield and a textarea field as validator.