accounting: fix accounting period constraint with node acquisition on invoices
We have a constraint preventing closing accounting periods if there are
still some accounting transactions that are in "current states" (ie. not
delivered / cancelled), but this constraint should not be fooled by
accounting lines in stock table that does not have an account as node,
but just an acquired organisation.