all: Do not pass query=None to catalog.
"query" argument is not supposed to be different from any other
argument (argument name is [pseudo-]column name, unknown coluns are
rejected, and "None" means "IS NULL" SQL comparison), but is accidentally
special-cased in ERP5Catalog. This special-casing is to be removed, so
update all callers relying on this glitch.
Also, avoid using Query.