ERP5Catalog: Do not retrieve workflow catalog variables.
These are used nowhere in the code. In many workflows, these
"catalog-available" variables are only configured so they appear as fields on
worklist definitions. So it is understood as more of an informative setting
having a local effect than a normative one having an effect on how
indexation happens.
So shave off a bit of useless time spent in indexation.
Also, it should be mentionned that when profiling this code causes a big
red-herring: this code is the first place where workflow history gets
loaded (and presumably indexed document itself), which draws attention
away from potentially more interesting areas.