Allow to stop draft test results
This is accompanying commit to "erp5_test_result: stop affecting last
remaining tests to all test nodes", where alarms redrafts running tests
for long running nodes.

In case of very long running test nodes, to which support is introduced
in this work, it is acceptable situation that test will get redrafted
and shall be stopped.