erp5_upgrader: pass activate_kw to ActiveProcess's constructor
FullUpgrade Alarm relies on getLastActiveProcess to return results, which makes
use of the catalog to return a result. If the Active Processes spawned by the
3 Alarms composing the upgrader (pre, upgrade, post) are not indexed before
the FullUpgrade Alarm collects the results, then the wrong active result is
returned, which then give the wrong feedback to the user.

It seems that the "defaultActivateParameterDict" method cannot overwrite
the "tag" parameter of indexation activities (even if it was working for
after_tag). This is why this commit is needed.