accounting: make P&L compatible with "british" accounting style
This report has always been based on "french layout", where there is first
all expenses on one page and all incomes on next page.

In the "british layout", there is only one sheet, showing income -
expense on the same layout ( as we can see today on
 https://en.wikipedia.org/wiki/Income_statement#Sample_income_statement
)

The idea is that we can configure financial section either with two
hierarchies for expense and income to have the french layout, or put all
the "british layout" hierarchy in income and leave expense empty.