erp5_post: Add HTML Post and Post Module.
Post is a comtainer which contains raw data without any metadata.
A post can be seen as any item send or received by an instant message application.
HTML Post is one type of post, which contains richtext in HTML and UTF-8.
Eugene Shen wrote a post system before, which inpired us. This bt5 is a
new implementation which under Tristan Cavelier's guide.