[erp5_http_proxy] Prototype of an usable HTTP proxy.
It can be used with an external method by providing the URL to proxy as a query parameter:
ERP5Site_getHTTPResource?url=URL

The query is synchronous (no portal_activities), so, the Zope thread will be blocked until the external request is other.
In order to prevent consuming all threads, there is a really aggressive timeout of 1 second only.
This is enough for a RSS reader.
If the timeout is to short, do not increase it, but change the proxy design to use portal_activities instead.

For now, only authenticated user can use it (to prevent becoming an open proxy).
Only GET query are allowed (until a use case required any other HTTP method).