erp5_data_notebook: Implement basic execution
By providing simple scripts and following the principle of storing code
locally on the CDNL it is easy to postExecution from CDN and simply
return URL to the document which will contain the result and then fetch
the result.

Use External Method, as more analysis of executed code will be needed.

Tests proves, that the call is in the end executed in restricted
environment, thanks to using PyData Script internally.

Accompanying test exposes API.

By using frame information, the internals of the called script are available
to outer caller.