erp5_officejs_support_request_ui: Use ECharts wrapper instead use Chartjs directly.
Chartjs can not work well. Add a wrapper for ECharts in a seperate MR. Use the wrapper to render chart. And also add a listbox to response the click event in chart. When user click an item in the chart. The listbox only show the items which related to the chart item.