CatalogTool: Split getCategoryParameterDict in 2
One part working on {base_category_id: [document_value_set]}, and the other
producing that structure out of a flat list of (base_category_id-prefixed)
category relative urls.
The former used to be included in the latter, but in some case caller
already has the documents handy instead of only urls, in which case we can
avoid a layer of document lookups.