BusinessTemplate: Maintain consisitency while adding values in modified_object_list
In al the TemplateItem objects, we consider the values being added to modified_object_list
in the format {<object_path>: [List of changed action]}. But, for ObjectTemplateItem we
used to add tuples instead of list. This was not getting tested as we always had cases
where we had old_xml and new_xml similar to each other.
This also explains why some tests were expecting tuples and not lists in this commit:
https://lab.nexedi.com/nexedi/erp5/commit/727477a32b1b5dafb14e01c28d07dd029fb4fac8

But, fortunately this came to notice after migrating SQL Method objects to behave as
ERP5fied objects.