BusinessTemplate: Reset portal_components on the fly while installing components
Before:

  self.portal_components.reset(force=True,
                               reset_portal_type_at_transaction_boundary=True)

After:

  self.portal_components.reset(force=True)

If the reset_portal_type_at_transaction_boundary is True, it does reset
component only once at the end of transaction.
(In detail, it reset component document module on the fly, and reset
 the component portal_types in the end of the transaction)

However, it is possible that those components are required in the middle of
the transaction while installing the business templates. For exmaple:
 - A method of a component is triggered while installing
 - A document component is required in a different business template,
   and those buisiness templates are installed inside a single transaction
   by upgrader.

Thus reset here on the fly.