[ui_test] Order (priority) of actions matter - first action becomes the default view of an object
Fixing testCloneObject unit test
The problem was that upon installation the generic list view became the default view of Foo Line thus was missing actions like Clone. This one is particularly hard to find because diff does not show this problem and moreover newly created action on a Portal Type is placed last by default. Thus this is a random error.

/reviewed-on https://lab.nexedi.com/nexedi/erp5/merge_requests/375