data protection: translate workflow messages when displayed
If we call Base_translateString before storing the message in the
history, the message will be translated when this action is used and
stored as translated using the language of the user executing this action.

The proper way is to store a message in history, so that it is rendered
when displayed.