SimulationTool: Remove input/output perimeter definition.
As per Jérome, who implemented the test, it was written to test the
current state rather than testing the desired outcome. And it makes
little sense to have (and test for) 100 being present in both debit and
credit columns ("normal" lines), and 0 to be present in the stat line.

Update test to check for a more consistent outcome.

Acked-by: Jérome Perrin <jerome@nexedi.com>