ERP5Type.ERP5Type.constructInstance: Update local roles after edit.
Local roles may have to be updated according to the properties set to object by init script and _edit. Without it, I have to write code to call updateLocalRolesOnSecurityGroups every time when creating an object, it is inconvenient and it is a big reason for proxy role utilisation so far.