accounting: allow any transaction type as causality of invoice
The main motivation is to allow to use Internal Invoice Transaction as
causality of Internal Invoice Transaction, so that when they are posted
the automatic grouping takes place.

A side effect is that now users can also create transaction as causality
of Balance Transaction, not really expected, but not really a problem.