ERP5Type.ERP5Type: Remove user_name parameter from updateLocalRolesOnDocument.
To pass a user name to give Owner local role does not make sense.
If Owner local role has to be given to a certain user, then it should
be done through Role Information or Role Definition. Also, when user
name is not given, it does not make sense to give the current user
Owner local role. Thus getSecurityManager().getUser().getId() is
removed.