ERP5LoginUserManager: Check Assignment is only meaningfull for Persons
The Assigments are only relevant if the User is based on Person Document, on SlapOS, an user come from a Computer or a Software Instance Document. This change makes the document compatible with other kinds of Users, as Assigment is overkill for Non-Person cases, as the presence of ERP5 Login is enough to determinate the User existence. 

The other alternative would be re-factorize to not duplicate code, which seems not the intention of how this code was written. 

/cc @jerome, @kazuhiko 



/reviewed-on https://lab.nexedi.com/nexedi/erp5/merge_requests/458