erp5.util.testnode: Add periodic test functionality
Periodic test functionality makes it possible to require that given test
will be run once per given period.

It is configured by server. The URL is in style:

  `periodic-prefix://<periodicity>`

where <periodicity> is time.strftime acceptable string.

The Buildout Section ID can be used to name the repository, and tested
revision sent back to the server.

For example in order to have tests be run at least once per day the
configuration on server is:

 * url: periodic-prefix://%Y/%m/%d
 * buildout section ID: daily