erp5_forge: Automatically import a zip package into erp5.
This commit mainly serve as an example.

Suppose we need to update the ckeditor. We can download the package zipfile.
Then upload the zipfile in ForgeModule_importZipFileDialog. The script will
automatically extract the files in it. And update or create(if not exist) files in erp5.

Please read the code and the comment in ForgeModule_importPackageFromZipFile.py
before using it. And add your modification if needed.