simulation: add a solver allowing to move quantity to another existing delivery
Usually, the split and defer dialog suggest you to define start and stop date. So if you have
already an existing delivery and you would like it to be updated, you have to set correct start
and stop date and, if all categories are fully the same, you might have some chance to see
this delivery updated.

To improve the case where you would like to move quantity to existing delivery, add option in the
solve divergence dialog of "Move to existing Delivery" that allows you to select similar deliveries. Then
this specific solver would set correct properties for you to make sure the expected delivery would be updated.