erp5_base: add Organisation's registration code and VAT code in their searchable text
It is a common (CRM principally) use case to search for Organisations using official identifying unique codes, as the registration code or the VAT code.

There is currently no way in ERP5 to look for entities using these unique codes, as they are not-generic enough to be indexed. For the same reason, creating a new table(s) to index Entities' identifiers may give something too blured to reach a consensus on. In my opinion fulltext search is a good solution for this case, and it shouldn't create compability issues for current projects. 

If it is thought necessary, I can also add the Social Security Code (for Organisations/Persons) in the searchable text properties, but it is a more complicated topic as it isn't a public information.

@nexedi 's project managers, what are your thoughts on it ?

/reviewed-on https://lab.nexedi.com/nexedi/erp5/merge_requests/463