item: Directly sort the output of Item_getTrackingList by date
Some listboxs using it were using default_sort to inject this
automatically, but it's better to do this directly in the script, this
way the script can be reused.

I'm not really convinced the script should honor sort_on parameter
received in kw, because misconfigured listbox can inject some parameters
that would cause more table to be joined.