AccountingTransactionModule_get*AgedBalanceLineList: Rework.
Make it more easily extensible by exposing a callback.
Stop loading documents in the simple version that are only needed by the
detailed version. Also, stop checking for this inside the main loop, and
instead use the callback mechanism.
Use a single query to look for mirror section title instead of one per
mirror section.
Behaviour change: mirror section (ie, the entity) title is not
translated anymore. It used to be because there did not seem to be a
reasion not to. It is not anymore to avoid loading rach entity from ZODB
in the default report. This rework makes customisation easier, including
overwriting this untranslated title with any desired transformation.