catalog tool: move compilation of python script inside changeObjectClass
Python script needs to be recompiled when they comes from a change of
class. This was done when business template was installing scripts, but
not when automatic migration from old catalog to new catalog was done.
So make sure to call _compile directly in changeObjectClass to make sure
Python Scripts would always work after a change of class.