erp5_crm: initial commit of the mime_post_module, and one of its sub-objects, the Internet Mail Message
The purpose of the Mime Post Module is to track what goes out and in ERP5, without any processing.
For exemple, the Internet Mail Message is an object wrapping Email envelopes. Then the envelope can be processed, its content will be used to create a Mail Message, its attachment transformed into documents,  and its headers readen to add source/destination to the Mail Message.