administration: fix incremental consistency check for non UTC timezone
When zope does not run in a UTC timezone,
last_active_process.getStartDate will be in local timezone.

If we just cast this to string and compare it against
indexation_timestamp which is UTC in our case (because we don't use SET
time_zone statement), we'll convert two dates in different timezone,
which practically in this case can lead to not checking some documents
or checking documents twice.