administration: fix incremental consistency check for non UTC timezone
last_active_process.getStartDate() is in Zope timezone, we should not
just pass a string representation to catalog but pass a DateTime
instance instead, to let catalog convert the date to UTC (because
catalog tables have UTC dates).

Where it gets more complicated is that unlike all DATETIME columns in
catalog that are UTC, indexation_timestamp is a TIMESTAMP column, which
depend on MySQL timezone settings. This also depend assume that MySQL is
configured with an UTC timezone.