administration: workaround alarm.getLastActiveProcess not during alarm execution
At this point, the alarm is active, but active process for this activity
was not created yet. By passing include_active, the latest active
process will be returned (not the previous one)

Honnestly, not sure if this falls into workaround or non documented
normal use case for include_active=True.