testnode: use PATH instead of command-line options to provide extra binaries
The /bin/ folder of all extra SR provided by the testnode are added to PATH.
SR must fixed either by moving executables or adding symlinks.

The testnode will stop looking at the parameter list of runTestSuite,
and --firefox_bin/--xvfb_bin are deprecated.

This is part of defining a generic and unique API between the testnode and
the tested software.