all: Remove recursiveImmediateReindexObject, and references to it.
This method does not belong to Folder (nor Base, even as an alias), as
it recurses wihout any limit in a tree whose node count is unknown.
Such generic recursion should either be deferred (split in multiple
transactions, each processing a limited chunk of the tree) or not be
recursive (which of course depends on the need).
There is a single caller left (constructInstance) where the subtree size
is assumed to be reasonable, as it was constructed inside current
transaction. So essentially rename this method so it is only used in such
conditions.