Base: Add isSubtreeIndexable and isAncestryIndexable
isAncestryIndexable is a parent-recursive version of isIndexable: it
will return False if any parent is non-indeable. This simplifies
disabling indexation of an entire tree, from its root to its deepest
leaves.
isSubtreeIndexable allows preventing any subobject from being indexable,
without having to mark each individual subobject as non-indexable. This
is different from isAncestryIndexable, as it allows having an indexable
root, but no children indexable. This is especially useful for Trash Bins,
where the bin itself should be indexable (because there is no reason not to)
but its content must not be indexed, while being left as it was found in its
original location.