Revert "erp5_officejs_support_request, erp5_post: Pass follow-up by value"
This reverts commit c49cf8d83501bcd22253d7e8a69f5bc0e93c51ef.
Callee is directly accessed from other places, so changing this API is
wrong.
This it not immediately visible in tests, but this makes
erp5_officejs_support_request_ui_test:testFunctionalSupportRequest
testCheckComment test randomly fail (apparently because the javascript
error message may appear after or before the test final step is
satisfied - which means this tests is bugged, but that's a separate
issue).