mrp: enable possibility to define Transformed Resource Lines with industrial phases
As part of a manufacturing, we can have products partially built (thus having a particular
industrial phase). However current code prevent setting as part of transformation lines a
resource associated to a particular industrial phase. There should be no reaons to avoid this.

industiral_phase is added to amount to allow getGeneratedAmountList (which is using
temp Amount documents) to handle correctly industrial_phase