[erp5_web_renderjs_ui]: Check for portal_status_message for errors
In case there is an error and the target response if of type json, we
should check for the portal_status_message and display it.

[erp5_hal_json_style]:
Do not set HTTP response code for error in Base_renderForm.
Add new script Base_returnFailureMessage to return response header for HTTP code errors.
Also, Remove useless message trunacation