deferred_style: factorize request argument filtering
This code was slightly different in three places:
 * Base_activateSimpleView: only use request.form.items()
 * Base_activateReport: use request.items()
 * Base_computeReportSection: use request.items()

request.items returns all items from request.other, request.form and
request.cookies.
In all cases, we are interested in request.form, so that report runs
with the same user input in request. We want to restore restore
request.other for Base_computeReportSection , but it does not harm in
other cases.
We always drop cookies to fix  #20180220-1DBF516 (as a consequence the
request will not strictly be identical, but rendering reports should not
depend on cookies)