ERP5Type: Recursively reindex created documents.
The rationale behind this change is to cover the case of sub-documents
modified as part of acquired setters (ex: subordination on Person which
actually sets it on a Career subdocument, or address on Person, etc).
As these value can be passed to newContent (and passed on to
constructInstance) along with immediate reindexation request, caller may
expect these to be also indexed.