erp5_deferred_style: Allow passing previous_skin_selection
To be used so that it can work in new UI

erp5_hal_json_style: Set previous_skin_selection in kw when Base_activateSimpleView

is to be called for deferred report

hal always use simple script to defer rendering

Simple script is able to render a valid response in case of missing email for example.
report_view can't, because it does not return HTML.
maybe report_view should be converted to a python script?

deferred support direct URL call

/Base_viewHistory?deferred_portal_skin=ODS&portal_skin=Deferred

deferred: mark error messages