testCMFCategory: update comment about categories membership to self
Add a warning about the doubt regarding the example in comment, to me
this does not depend on category being member of self.

One reason for the membership to self - in my understanding - is to make
inventory API work the same with movement to meta node categories or
movement to nodes member of this category when using
getInventory(node_category=(the meta node category))

Also remove comment about what's said to be a "mistake" ( the svn url should
be equivalent to git commit eedb2b294b7d06e531b9267160d2fb9a4b436df5 )
because this was never reverted and does not seem to be a mistake, the
purpose of the change is to prevent having categories member of
themselves twice.