fixup! Folder: Make recursiveReindexObject scalable by calling _recurseCallMethod
With commit babbf56f9b2bf266e5722be4b881868bb01fc362, some objects that were
not indexed are now indexed, and this fails when these objects can't be indexed.

An exemple is when a PythonScript is stored inside a Web Site:

  Traceback (innermost last):
  Module Products.CMFActivity.ActivityTool, line 1360, in invokeGroup
    traverse(method_id)(expanded_object_list)
  Module Products.ERP5Catalog.CatalogTool, line 939, in catalogObjectList
    super(CatalogTool, self).catalogObjectList(tmp_object_list, **m.kw)
  Module Products.ZSQLCatalog.ZSQLCatalog, line 828, in catalogObjectList
    catalog_value=catalog,
  Module Products.ERP5Catalog.CatalogTool, line 875, in wrapObjectList
    - __traceback_info__: (<PythonScript at /erp5/web_site_module/...>,)
    and document_object._getAcquireLocalRoles():
  AttributeError: _getAcquireLocalRoles

Before, such object was skipped because it does not have recursiveReindexObject.