ERP5TypeTestCase: rework user / passord generation
Manager username and password are now available as class attributes.

Many tests have been updated to not generate manager user and use the
existing one instead. When tests need to create users (for example
users in the root acl_users), we try to delete the users afterward.

Another significant changes is that we no longer have the same user in
ERP5/acl_users and /acl_users, some tests were logging in as the root
user in ways that never seemed intentional.

ERP5TypeFunctionalTestCase was updated to use self.manager_username and
manager_password instead of hardcoding ERP5TypeTestCase: and also some
cleanups, the HTTP server is already started at this point and host/port
were not used.