ERP5: implement Telephone.asURL for non-detailed coordinates
Telephone.asUrl (which is used in erp5_short_message) properly supported
detailed coordinates, but had no support for coordinate text.

In that case, we don't try to be to clever and just use the coordinate
(almost) as is, just using the tel: prefix and removing spaces which are
not legal.